+++ fe/src/Components/Wallet/Coins.jsx
@@ -1,4 +1,4 @@
-import { SetAlertMessage } from "App";
+import { SetAlertMessage } from "Context/AlertMessageProvider";
 import useInvestment from "Hooks/useInvestment";
 import useWallet from "Hooks/useWallet";
 import { INIT_ALERT_MESSAGE } from "Helper/constant";
